requirements.txt
Django==1.9.5
django-hstore==1.4.2
django-mptt==0.8.3
lxml==3.6.0
psycopg2==2.6.1
